+++ CHANGELOG.md
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 None
 
+## 0.2.9 - 2026-01-08
+
+### Fixed
+
+fix: prevent visual artifacts when switching to shorter content
+  Use the new `reset()` function instead of creating again a new code editor !
+fix: prevent mouse to capture events when out of bounds
+
+
 ## 0.2.8 - 2026-01-08
 
 ### Fixed